+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Modify the arguments to narrow down the import target with `--dbt-model` (#532)
 
 
 ## [0.10.15] - 2024-10-26