+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Modify the arguments to narrow down the import target with `--dbt-model` (#532)
-
+- SodaCL: Prevent `KeyError: 'fail'` from happening when testing with SodaCL
 
 ## [0.10.15] - 2024-10-26