+++ CHANGELOG.md
@@ -11,10 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Type conversion when importing contracts into dbt and exporting contracts from dbt (#534)
 
 ### Fixed
 - Modify the arguments to narrow down the import target with `--dbt-model` (#532)
-
+- SodaCL: Prevent `KeyError: 'fail'` from happening when testing with SodaCL
 
 ## [0.10.15] - 2024-10-26